File IO Lib update : 2.6.11
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libhxcfe/trunk@1410 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ libhxcfe/sources/thirdpartylibs/FATIOlib/example.c
@@ -7,9 +7,9 @@ int media_init()
     return 1;
 }
 
-int media_read(uint32_t sector, unsigned char *buffer, uint32_t sector_count)
+int media_read(unsigned long sector, unsigned char *buffer, unsigned long sector_count)
 {
-    uint32_t i;
+    unsigned long i;
 
     for (i=0;i<sector_count;i++)
     {
@@ -24,9 +24,9 @@ int media_read(uint32_t sector, unsigned char *buffer, uint32_t sector_count)
     return 1;
 }
 
-int media_write(uint32_t sector, unsigned char *buffer, uint32_t sector_count)
+int media_write(unsigned long sector, unsigned char *buffer, unsigned long sector_count)
 {
-    uint32_t i;
+    unsigned long i;
 
     for (i=0;i<sector_count;i++)
     {
